Handle multiple datasets in data entry
