+++ src/components/dataentry/DataEntrySelectionPage.js
@@ -15,16 +15,7 @@ import { withRouter } from "react-router";
 import PeriodPicker from "../shared/PeriodPicker";
 import PortalHeader from "../shared/PortalHeader";
 import LinkedContract from "./LinkedContract";
-
-function isDataSetComplete(completeDataSetRegistration) {
-  if (completeDataSetRegistration == undefined) {
-    return false;
-  }
-  if ("completed" in completeDataSetRegistration) {
-    return completeDataSetRegistration.completed;
-  }
-  return true;
-}
+import { buildFormData } from "./forms";
 
 const DataEntrySelectionPage = ({ history, match, periodFormat, dhis2 }) => {
   const { t, i18n } = useTranslation();
@@ -97,223 +88,14 @@ const DataEntrySelectionPage = ({ history, match, periodFormat, dhis2 }) => {
 
         const api = await dhis2.api();
         if (match.params.dataEntryCode) {
-          const dataEntry = dataEntryRegistry.getDataEntry(match.params.dataEntryCode);
-          const dataSet = dataEntry.dataSetId
-            ? await api.get("/dataSets/" + dataEntry.dataSetId, {
-                fields:
-                  "id,name,periodType,access,dataSetElements[dataElement[id,name,valueType,optionSet[options[code,name]],categoryCombo[id,name,categoryOptionCombos[id,name]]]]",
-              })
-            : { dataSetElements: [] };
-
-          const dataElementsById = _.keyBy(
-            dataSet.dataSetElements.map((dse) => dse.dataElement),
-            (de) => de.id,
-          );
-
-          let completeDataSetRegistration;
-          const dsc = dataEntry.dataSetId
-            ? await api.get("completeDataSetRegistrations", {
-                dataSet: dataEntry.dataSetId,
-                period: period,
-                orgUnit: activeContract.orgUnit.id,
-              })
-            : { completeDataSetRegistrations: [] };
-
-          completeDataSetRegistration = dsc.completeDataSetRegistrations
-            ? dsc.completeDataSetRegistrations[0]
-            : undefined;
-          const dv = dataEntry.dataSetId
-            ? await api.get("/dataValueSets", {
-                dataSet: dataEntry.dataSetId,
-                orgUnit: activeContract.orgUnit.id,
-                period: period,
-              })
-            : { dataValues: [] };
-          let rawValues = dv.dataValues || [];
-          if (dataEntryRegistry.fetchExtraData) {
-            const extraValues = await dataEntryRegistry.fetchExtraData(api, activeContract.orgUnit, period, dataEntry);
-            rawValues = rawValues.concat(extraValues);
-
-            const extraDataElements = await dataEntryRegistry.fetchExtraMetaData(
-              api,
-              activeContract.orgUnit,
-              period,
-              dataEntry,
-            );
-
-            for (let extraDe of extraDataElements) {
-              dataElementsById[extraDe.id] = extraDe;
-            }
-          }
-          const defaultCoc = (
-            await api.get("categoryOptionCombos", {
-              filter: "name:eq:default",
-            })
-          ).categoryOptionCombos[0].id;
-
-          const indexedValues = _.groupBy(rawValues, (v) =>
-            [v.orgUnit, v.period, v.dataElement, v.categoryOptionCombo].join("-"),
-          );
-
-          let calculator;
-          if (dataEntryRegistry.getCalculator) {
-            calculator = dataEntryRegistry.getCalculator(activeContract.orgUnit, period, match.params.dataEntryCode);
-            if (calculator) {
-              calculator.setIndexedValues(indexedValues);
-              calculator.setDefaultCoc(defaultCoc);
-            }
-          }
-
-          const newFormData = {
-            period,
-            values: rawValues,
-            indexedValues: indexedValues,
-            orgUnit: activeContract.orgUnit,
-            dataSetComplete: isDataSetComplete(completeDataSetRegistration),
-            completeDataSetRegistration: completeDataSetRegistration,
-            dataSet: dataSet,
-            dataElementsById: dataElementsById,
-            calculator: calculator,
-            valids: {},
-            errors: {},
-            updating: {},
-
-            isDataSetComplete() {
-              return this.dataSetComplete;
-            },
-            isDataWritable() {
-              return this.dataSet && this.dataSet.access && this.dataSet.access.data.write;
-            },
-            error(de) {
-              return this.errors[this.getKey(de)];
-            },
-            getKey(de) {
-              const deCoc = de.split(".");
-              return [activeContract.orgUnit.id, period, deCoc[0], deCoc[1] || defaultCoc].join("-");
-            },
-            getError(de) {
-              const key = this.getKey(de);
-              return this.errors[key];
-            },
-            isModified(de) {
-              const key = this.getKey(de);
-              return this.valids[key] == true;
-            },
-            isUpdating(de) {
-              const key = this.getKey(de);
-              return this.updating[key] == true;
-            },
-            isInvalid(de) {
-              const key = this.getKey(de);
-              return this.valids[key] == false;
-            },
-            getValue(de) {
-              const key = this.getKey(de);
-              const ourValues = this.indexedValues[key];
-              return ourValues ? ourValues[0] : undefined;
-            },
-            getCalculatedValue(hesabuPackage, formulaCode, period, orgUnit, activity) {
-              let orgUnitId = orgUnit ? orgUnit.id : activeContract.orgUnit.id;
-              const calculatorFunction = activity
-                ? `${hesabuPackage.code}_${activity.code}_${formulaCode}_${orgUnitId}_${period}`
-                : `${hesabuPackage.code}_${formulaCode}_${orgUnitId}_${period}`;
-              if (this.calculator && this.calculator[calculatorFunction]) {
-                return this.calculator[calculatorFunction]();
-              }
-            },
-            async updateValue(de, value) {
-              const deCoc = de.split(".");
-              const key = this.getKey(de);
-              if (this.updating[key]) {
-                return;
-              } else {
-                this.updating[key] = true;
-              }
-              const newValue = {
-                de: deCoc[0],
-                co: deCoc[1] || defaultCoc,
-                ds: dataSet.id,
-                ou: activeContract.orgUnit.id,
-                pe: period,
-                value: value,
-              };
-              try {
-                await dhis2.setDataValue(newValue);
-                let newIndexedValues = this.indexedValues;
-                if (this.indexedValues[key]) {
-                  newIndexedValues[key] = [{ ...this.indexedValues[key][0], value: value }];
-                } else {
-                  newIndexedValues[key] = [{ dataElement: newValue.de, value: value }];
-                }
-                if (calculator) {
-                  calculator.setIndexedValues(newIndexedValues);
-                }
-                this.valids[key] = true;
-                this.errors[key] = undefined;
-                this.updating[key] = false;
-
-                setFormData({ ...this });
-              } catch (error) {
-                this.valids[key] = false;
-                this.errors[key] = error.message;
-                this.updating[key] = false;
-
-                setFormData({ ...this });
-              }
-            },
-            async toggleComplete(calculations) {
-              if (completeDataSetRegistration && "completed" in completeDataSetRegistration) {
-                // newer dhis2 version just toggle "completed" or create one
-                await api.post("completeDataSetRegistrations", {
-                  completeDataSetRegistrations: [
-                    {
-                      dataSet: this.dataSet.id,
-                      period: this.period,
-                      organisationUnit: activeContract.orgUnit.id,
-                      completed: !this.dataSetComplete,
-                    },
-                  ],
-                });
-              } else if (completeDataSetRegistration && this.dataSetComplete) {
-                // older dhis2 delete the existing completion record
-                await api.delete(
-                  "completeDataSetRegistrations?ds=" +
-                    this.dataSet.id +
-                    "&pe=" +
-                    this.period +
-                    "&ou=" +
-                    activeContract.orgUnit.id +
-                    "&multiOu=false",
-                );
-              } else {
-                // older dis2 delete the existing completion record
-
-                await api.post("completeDataSetRegistrations", {
-                  completeDataSetRegistrations: [
-                    {
-                      dataSet: this.dataSet.id,
-                      period: this.period,
-                      organisationUnit: activeContract.orgUnit.id,
-                      completed: !this.dataSetComplete,
-                    },
-                  ],
-                });
-              }
-
-              if (calculations && !this.dataSetComplete) {
-                const orbf2 = PluginRegistry.extension("invoices.hesabu");
-                calculations.forEach((calculation) => orbf2.calculate(calculation));
-              }
-
-              const updatedFormaData = {
-                ...this,
-                dataSetComplete: !this.dataSetComplete,
-              };
-
-              setFormData(updatedFormaData);
-            },
-          };
-
+          const newFormData = await buildFormData({
+            api: api,
+            dataEntryCode: match.params.dataEntryCode,
+            activeContract: activeContract,
+            dataEntryRegistry: dataEntryRegistry,
+            period: period,
+            setFormData: setFormData,
+          });
           setFormData(newFormData);
         }
       } catch (error) {
